+++ ecj/ec/app/regression/Regression.java
@@ -193,7 +193,7 @@ public class Regression extends GPProblem implements SimpleProblemForm
                 
             // the fitness better be KozaFitness!
             KozaFitness f = ((KozaFitness)ind.fitness);
-            f.setStandardizedFitness(state,(float)sum);
+            f.setStandardizedFitness(state, sum);
             f.hits = hits;
             ind.evaluated = true;
             }
